feat: add trackedFunction and function-resource
trackedFunction is powered by an *also* new `resource` function (exported from
uti/function-resource).

`trackedFunction` now has additional state information, such as
- isLoading
- isPending
- isError
- isResolved
- error
- value (pre-existing)

To use the function-resource, you'll need the @use decorator

This is a different `@use` from before, and is why it's exported from
the same import path as the function-resource.

There was pretty much no usage of the old `@use`, so I think this change
is ok.