feat(function-resource): support encapsulated tracked state
This is based on ideas from https://wycats.github.io/polaris-sketchwork/reactivity.html
Where a function-based resource _may_ return an arrow function instead
of a non-function-value. This allows the function-based resource to
conusme and update its own tracked state without invalidating itself.

Previously, in order for function-based resources to have their own
tracked state, an object had to be returned, and it was up to the
consumer to consume that tracked state via property access.

With this arrow-function return, we can provide a single reactive value
via the resource as well as encapsulated state.