fix(Resource): class-based resource had no inference on the thunk
previously, when typing out the return object of the thunk when using a
class-based resource, the whole object would be "unknown", which means
that you could provide _anything_ there, and it wouldn't matter to the
type checker. Now, you'll get intellisense in the thunk as well as
errors when your thunk is incorrect. This may cause errors during the
upgrade to this version, but it should be a "correct change" to resolve.