Fix polling helper example in CHANGELOG
The example in the original had some bugs. As it initiated a new `setInterval` on every compute it just keep adding every time Ember recomputed the helper. It also would register a destructor on every compute.

This change uses a WeakMap to track the intervals and only registers a destructor once.